Handle diagnostics in the DepGraph.
